Switch ID generation to UUID, add collision error type
Rather than using the 64-byte length random ID that we were using (32
bytes represented by hex), switched IDs to be standard v4 random UUIDs.
This still offers a near impossible risk of collisions, with the sheer
amount of records required to approach any iota of a probability of
hitting one being so high that you are probably going to hit other
scalability issues with this store before that. Way before that.

Nonetheless, also added a IDCollisionError type so that one can check
for collisions via a type switch, if it's ever necessary.