example: Add gRPC subscriber example
Added a gRPC subscriber example, allowing a gRPC server to dispatch
multiple subscription requests to a single stream. This could probably
be extended to a publisher as well.

There is a minor issue right now where currently multiple subscribers
are set up for a single stream, which might not scale that well. Further
work on this gRPC example could ensure that a single FS watcher is
multiplexed to any number of subscriber endpoints.