sub: Add DumpSorted, DumpSortedReverse, and IndexedEvent
This allows for ascending and descending sort of the data returned by
Dump. The sort criteria is defined in the event by implementing
IndexedEvent, which provides a Less method that resembles the one
required by the standard sort package's sort.Interface, except that i is
a receiver and not an index. Examples are in the docs.