sub: More tests
Added for SubscribeCallback and Dump.

Have also identified a race with Subscribe that needs to be fixed - more
than likely due to delays in the inotify watch being fully set up.  The
current change with synching up the event publish to get it right before
the sub is the best I could do right now - incidentally, trying to
synchronously send the event after the sub (by wrapping the sub in a
goroutine) just made the race worse.

Will probably refactor Subscribe so that it asserts that can see
IN_CLOSE_WRITE before returning. This will make Subscribe non-blocking,
so another waiter function will be added that performs most of the
waiting tasks currently in Subscribe.