sub: Refactor Subscriber for new setup/watch logic
Subscribe is being refactored so that it returns on creation. Errors are
now piped through an error channel instead and returned to an Error()
function similar to the architecture of a context, and the done channel
is now the source of truth of when the stream has ended, error or
otherwise.

This corrects some races that were happening in testing due to the
publish being run before the subscribe, albeit as close to before as
possible. A benchmark test has also been added for the entire watcher
suite to ensure that no races are currently happening.

More refactoring will need to be done, namely to SubscribeCallback, and
tests for trapping errors on Subscribe will need to be added too, now
that the watch goroutine has been decoupled.