store: New package, common stream initialization functionality
I'm starting work to de-couple most of the non-streaming subscription
stuff from the subscriber. This means that the Fetch and Dump functions
will need the ability to initialize stream without the assistance of
Subscriber.

The first part of this work is to de-couple the common workload of
initializing the stream directory. This has been moved to a new store
package.