sub: Finish refactoring Subscriber and Dump/Fetch
Dump and Fetch are now de-coupled from the subscriber. The API for
Subscriber has change drastically, and the README and top-level GoDoc
will need to be updated accordingly. Complete coverage needs to be added
for the sub package as well, in addition to re-org of the tests in pub
so that there is coverage for the store package as well (most of the
heavy lifting in initializing a dir is now done by NewStream).