sub: Remove NewSubscriberWithCallback
This function honestly does not bring much to the table, and in addition
to that introduces opportunity to create confusion and bad code, as
Queue is still exposed via the regular subscriber, meaning that someone
could easily put in a callback and then also still attempt to pull
events off of the exposed Queue channel. I'd rather delete it than test
it, so that's what I'm doing. :P

Instead, someone can hand-off to a function in the typical event loop
and still get the "callback" functionality, at the expense of a few
lines of boilerplate.